+++ resources/views/siggy/activities/dscan.php
@@ -1,0 +1,74 @@
+<div id="activity-dscan" class="wrapper" style="display:none">
+<p>
+	This page lists all systems that have signatures stored and the latest time they were entered as the Last Scan column.</a>
+</p>
+
+
+<div class="container-fluid">
+	<div class="row">
+		<div class="col-md-4">
+			<div class="panel panel-default">
+				<div class="panel-heading">Everything</div>
+				<table cellspacing='1' class='table table-striped'>
+					<tr>
+						<td>Enyo</td>
+					</tr>
+				</table>
+			</div>
+		</div>
+		<div class="col-md-4">
+			<div class="panel panel-default">
+				<div class="panel-heading">Ships</div>
+				<table cellspacing='1' class='table table-striped'>
+					<tr>
+						<td>Enyo</td>
+					</tr>
+				</table>
+			</div>
+			
+			<div class="panel panel-default">
+				<div class="panel-heading">Structures</div>
+				<table cellspacing='1' class='table table-striped'>
+					<tr>
+						<td>Astrahaus</td>
+					</tr>
+				</table>
+			</div>
+		</div>
+		<div class="col-md-4">
+		
+			<div class="panel panel-default">
+				<div class="panel-heading">Filters</div>
+				<div class='panel-body'>
+					<div class="btn-group" role="group" aria-label="...">
+						<button type="button" class="btn btn-default">Condensed</button>
+						<button type="button" class="btn btn-default">Uncondensed</button>
+					</div>
+					<br />
+					<div class="btn-group" role="group" aria-label="...">
+						<button type="button" class="btn btn-default">All</button>
+						<button type="button" class="btn btn-default">Ships</button>
+						<button type="button" class="btn btn-default">Structures</button>
+					</div>
+					<br />
+					<div class="btn-group" role="group" aria-label="...">
+						<button type="button" class="btn btn-default">All</button>
+						<button type="button" class="btn btn-default">On Grid</button>
+						<button type="button" class="btn btn-default">Off Grid</button>
+					</div>
+
+					<div>
+						<h5>Ships:</h5>
+						<ul>
+							<li>Total: </li>
+							<li>On Grid: </li>
+							<li>Off Grid: </li>
+						</ul>
+					</div>
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+
+</div>