+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ## Major Changes
 - Now 100% more leftpad dependency. Ported over frontend javascript to fancy pancy Typescript + webpack system from legacy old school javascript world
 - New router handling changes between activities (i.e. Scan vs. Thera page) which adds the pages to the browser history and provides links to the pages
+- DScan reimplemented and part of the main siggy application, url in browser can be shared with group members
+
+## Changes
+- Dscan intel section moved to new system matching POSes and structures
 
 ## Fixes
 - Fix double data fetch on load of siggy page. Wooo yay Typescript.